test: move empty string coverage into existing suites
Keep the regression coverage with existing sync and async host-function tests instead of a standalone file.
